refactor(traces): give traces a front door, and promote what it exposes (LFE-14804)
`index.ts` at the feature root — named re-exports only, and the list starts as
exactly what the eight external consumers already imported. Nothing added for a
future caller. The eight now import `@/src/features/traces`.

Routing through the index only moves rule 7 if the surface points at things
buried inside one component, so the four modules the outside world actually uses
are now feature-level, per the RFC's promote-to-a-sibling rule:

  IOPreview/hooks/useChatMLParser.ts          -> traces/hooks/
  IOPreview/components/ChatMessageList.tsx    -> traces/components/
  IOPreview/components/ThinkingBlock.tsx      -> traces/components/
  IOPreview/components/chat-message-utils.ts  -> traces/fns/chatMessageUtils.ts

Same for the one internal reach the sensor flagged: contexts/TraceDataContext
was importing TraceTimeline/timeline-calculations, now traces/fns/
timelineCalculations.ts.

rule 7: 13 -> 10, rule 8 (incoming): 9 -> 3, rule 9: 2 -> 1.

The index is client-safe on purpose: `server/` is absent, so importing it can
never pull Prisma or ClickHouse into a browser bundle. That leaves the three
server-side consumers of server/legacyIoSearch importing it directly — a
full-stack feature needs a second, server-side surface and the RFC has no rule
for one yet. Flagged, not faked.

Four modules needed their relative imports made absolute before they could move
at all; the tool refuses to edit a file it is moving. Those edits are the RFC's
own stated reason absolute imports keep moves clean.
