+++ web/src/features/annotation-queues/components/processors/SessionAnnotationProcessor.tsx
@@ -10,7 +10,7 @@ import { asCommentCounts } from "@/src/components/session/sessionDetailPageTypes
 import { useState, useMemo, useCallback } from "react";
 import { Button } from "@/src/components/ui/button";
 import { ItemBadge } from "@/src/components/ItemBadge";
-import { CopyIdsPopover } from "@/src/features/traces/components/CopyIdsPopover";
+import { CopyIdsPopover } from "@/src/features/traces";
 import { Badge } from "@/src/components/ui/badge";
 import { Separator } from "@/src/components/ui/separator";
 import Link from "next/link";
