+++ web/src/features/playground/page/components/GenerationOutput.tsx
@@ -5,7 +5,7 @@ import { ChatMessageRole, ChatMessageType } from "@langfuse/shared";
 import { BracesIcon, Check, Copy, Plus } from "lucide-react";
 import { ToolCallCard } from "@/src/components/ChatMessages/ToolCallCard";
 import { copyTextToClipboard } from "@/src/utils/clipboard";
-import { ThinkingBlock } from "@/src/features/traces/components/IOPreview/components/ThinkingBlock";
+import { ThinkingBlock } from "@/src/features/traces";
 
 export const GenerationOutput = () => {
   const [isCopied, setIsCopied] = useState(false);
