+++ web/src/features/traces/components/IOPreview/components/ChatMessage.tsx
@@ -20,8 +20,8 @@ import {
   parseToolCallsFromMessage,
   hasThinkingContent,
   hasRedactedThinkingContent,
-} from "./chat-message-utils";
-import { ThinkingBlock, RedactedThinkingBlock } from "./ThinkingBlock";
+} from "../../../fns/chatMessageUtils";
+import { ThinkingBlock, RedactedThinkingBlock } from "../../ThinkingBlock";
 import { type IOPreviewContentMode } from "../IOPreview";
 
 // View mode for pretty/json toggle
