+++ web/src/features/traces/components/IOPreview/components/ToolCallDefinitionCard.clienttest.tsx
@@ -5,7 +5,7 @@ import {
   ToolCallDefinitionCard,
   type ToolDefinition,
 } from "./ToolCallDefinitionCard";
-import type { ToolCallInvocation } from "../hooks/useChatMLParser";
+import type { ToolCallInvocation } from "../../../hooks/useChatMLParser";
 
 function createMemoryStorage(): Storage {
   const store = new Map<string, string>();
