+++ web/src/features/traces/components/IOPreview/components/ToolCallDefinitionCard.stories.tsx
@@ -3,7 +3,7 @@ import {
   ToolCallDefinitionCard,
   type ToolDefinition,
 } from "./ToolCallDefinitionCard";
-import type { ToolCallInvocation } from "../hooks/useChatMLParser";
+import type { ToolCallInvocation } from "../../../hooks/useChatMLParser";
 
 const baseTools: ToolDefinition[] = [
   {
