+++ web/src/features/traces/components/ThinkingBlock.tsx
