+++ web/src/features/traces/fns/chatMessageUtils.ts
@@ -1,7 +1,7 @@
 import type { z } from "zod";
 import type { ChatMlMessageSchema } from "@/src/components/schemas/ChatMlSchema";
 import type { combineInputOutputMessages } from "@/src/utils/chatml";
-import type { IOPreviewContentMode } from "../IOPreview";
+import type { IOPreviewContentMode } from "@/src/features/traces/components/IOPreview/IOPreview";
 
 export type ChatMlMessage = z.infer<typeof ChatMlMessageSchema>;
